Split server code across modules
